requirement.txt
Django==2.1.7
django-admin-rangefilter==0.3.12
python-dateutil==2.7.5
pytz==2018.9
six==1.12.0